+++ benches/filter_bench.rs
@@ -17,16 +17,16 @@ fn bench_filter(c: &mut Criterion) {
 
     let mut group = c.benchmark_group("filter");
 
-    let mut pt1_filter = Pt1Filterf32::new(1.0);
-    let mut pt2_filter = Pt2Filterf32::new(1.0);
-    let mut pt3_filter = Pt3Filterf32::new(1.0);
+    let mut pt1_filter = Pt1Filterf32::new();
+    let mut pt2_filter = Pt2Filterf32::new();
+    let mut pt3_filter = Pt3Filterf32::new();
     let mut biquad_filter = BiquadFilterf32::new();
     let mut median_filter3 = MedianFilter3f32::new();
     let mut median_filter5 = MedianFilter5f32::new();
     let mut ma_filter4 = MovingAverageFilter4f32::new();
-    let mut skew_limiter = SlewRateLimiterf32::new(10.0, 100.0, 0.1);
+    let mut skew_limiter = SlewRateLimiterf32::with_rates(10.0, 100.0, 0.1);
 
-    let mut pt1_v3_filter = Pt1FilterVector3df32::new(1.0);
+    let mut pt1_v3_filter = Pt1FilterVector3df32::new();
 
     _ = group.throughput(Throughput::Elements(1));