+++ CHANGELOG.md
@@ -23,6 +23,21 @@ Once the API has stabilized this project will adopt semantic versioning, the fir
 
 ### Security
 
+## [0.1.4] - 2026-05-16
+
+### Changed
+
+- Use vqm version 0.1.5
+- `new` and other constructors to be `const`.
+- Grouped `BiquadFilter` coefficients into a `struct`.
+- Improved documentation.
+
+## [0.1.3] - 2026-05-06
+
+### Changed
+
+- Use vqm version 0.1.3
+
 ## [0.1.2] - 2026-05-02
 
 ### Changed